Updated the code so that TripleA is again compatible with java 1.5. (The two lines that aren't compatible are within catch blocks and do not cause the user any errors)
To get Netbeans to create UI code that is compatible with 1.5, I had to tell it to use the Swing Layout Extensions Library instead of Standard Java 6.0 Code. This lets it run on the older jdk, but I think it also means that the swing layour extensions library will have to be packaged with TripleA now, I'm not too sure. (It added the library to the project's Libraries section within the project's properties window.)